Correct Style/IfUnlessModifier offense in jobs

+++ src/api/app/jobs/status_history_rescaler_job.rb
@@ -4,9 +4,7 @@ class StatusHistoryRescalerJob < ApplicationJob
   # this is called from a delayed job triggered by clockwork
   def perform
     maxtime = StatusHistory.maximum(:time)
-    if maxtime
-      StatusHistory.where('time < ?', maxtime - 365 * 24 * 3600).delete_all
-    end
+    StatusHistory.where('time < ?', maxtime - 365 * 24 * 3600).delete_all if maxtime
 
     keys = StatusHistory.distinct.pluck(:key)
     keys.each do |key|
